refactor(cli): organize app modules
