+++ kapsl-runtime/crates/kapsl-cli/src/app/cli/commands/mod.rs
@@ -1,0 +1,11 @@
+//! Schemas for commands that do not start the inference runtime.
+
+mod backends;
+mod extensions;
+mod models;
+mod packages;
+
+pub(crate) use backends::*;
+pub(crate) use extensions::*;
+pub(crate) use models::*;
+pub(crate) use packages::*;
