+++ kapsl-runtime/crates/kapsl-cli/src/app/config/mod.rs
@@ -1,0 +1,15 @@
+//! Resolution of CLI and environment input into runtime policy.
+
+use super::*;
+
+pub(crate) mod constants;
+mod model_loading;
+mod onnx_session;
+mod performance;
+mod runtime;
+
+pub(crate) use constants::*;
+pub(crate) use model_loading::*;
+pub(crate) use onnx_session::*;
+pub(crate) use performance::*;
+pub(crate) use runtime::*;
