+++ kapsl-runtime/crates/kapsl-cli/src/app/startup/mod.rs
@@ -1,0 +1,5 @@
+//! Runtime startup orchestration.
+
+mod bootstrap;
+
+pub(crate) use bootstrap::*;
