+++ kapsl-runtime/crates/kapsl-cli/src/app/terminal/mod.rs
@@ -1,0 +1,9 @@
+//! Terminal presentation and progress reporting.
+
+mod ansi;
+mod progress;
+mod status;
+
+pub(crate) use ansi::*;
+pub(crate) use progress::*;
+pub(crate) use status::*;
